Add change-password UI, Shift+Enter multiline messages, and fix hours-report crash
Adds a "שינוי סיסמה" entry in the sidebar wired to the change-password API.
Message composer and edit-message form are now textareas: Enter sends/saves,
Shift+Enter inserts a newline.

Fixes a crash in the admin hours report: Postgres returns aggregated
SUM(...) as a string, and the client called .toFixed() on it directly. Once
real time-clock entries existed this threw and broke the "שעות עבודה" tab.
Cast to float8 in the query so it comes back as a proper number.

+++ server/db.js
@@ -383,7 +383,7 @@ async function getHoursReport({ from, to, user_id, user_ids } = {}) {
   const { rows } = await query(
     `SELECT u.id AS user_id, u.username,
        COUNT(tl.id)::int AS entries,
-       SUM(EXTRACT(EPOCH FROM (tl.clock_out - tl.clock_in)) / 3600) AS total_hours
+       SUM(EXTRACT(EPOCH FROM (tl.clock_out - tl.clock_in)) / 3600)::float8 AS total_hours
      FROM time_logs tl
      JOIN users u ON u.id = tl.user_id
      ${where}
